% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{apply_binseg}
\alias{apply_binseg}
\title{Applies Binary Segmentation Algorithm to the given data data_vec}
\usage{
apply_binseg(data_vec, kmax)
}
\arguments{
\item{data_vec}{The 1 dimensional input chromosome data}

\item{kmax}{Maximum number of segments to divide.}
}
\value{
Status code of the algorithm.

  For description of how Binseg algorithm works, please see the following url.
  \url{https://arxiv.org/abs/1801.00718}
}
\description{
\code{apply_binseg} returns the status of a successful Binary Segmentation algorithm on the given parameters.
Most of the passed parameters are set for the result.
}
\examples{
apply_binseg(c(0.1, 0, 1, 1.1, 0.1, 0), 3)
}
