% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_dynprogseg.R
\name{apply_dynprogseg}
\alias{apply_dynprogseg}
\title{Applies Segmentation algorithm to the given data data.mat}
\usage{
apply_dynprogseg(data.dt, max.segments)
}
\arguments{
\item{data.dt}{A list of doubles representing the sequence data}

\item{max.segments}{Maximum number of segments to calculate}
}
\value{
A data.table of 3 columns with segments

  For description of how Binary segmentation and Dynamic Programming algorithm works in change point detection, please refer to this paper
  \url{https://arxiv.org/abs/1801.00718}
}
\description{
\code{Segmentation} returns the ...
}
\examples{
library(data.table)
apply_dynprogseg(iris[,1], 3)

}
