% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Segmentation-package.R
\docType{package}
\name{Segmentation}
\alias{Segmentation}
\title{Segmentation: A package for applying Segmentation algorithms}
\description{
The Segmentation package provides the following function:
apply_binseg
apply_dynprogseg
}
